made some progress; commands working

+++ internal/models/database.go
@@ -1,6 +1,6 @@
 package models
 
 type DbClient interface {
-	GetUserBalance(string) (float32, error)
-	SetUserBalance(string, float32) error
+	GetUserBalance(string) (float64, error)
+	SetUserBalance(string, float64) error
 }
